+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All Notable changes to `League\Uri` will be documented in this file
 ### Fixed
 
 - Adding `SensitiveParameter` attribute in the `Uri` and the `BaseUri` class.
+- Improve PSR-7 `Http` class implementation.
 
 ### Deprecated